main: Fix compilation warning
io-sensor-proxy.c: In function ‘process_scan’:
iio-sensor-proxy.c:790:3: warning: format ‘%d’ expects argument of type ‘int’, but argument 4 has type ‘ssize_t’ [-Wformat=]
    g_debug ("Not enough data to read (read_size: %d scan_size: %d)", data.read_size, (int) or_data->scan_size);
    ^

+++ iio-sensor-proxy.c
@@ -786,7 +786,7 @@ process_scan (SensorData data, OrientationData *or_data)
 	 * Just read the last one */
 	i = (data.read_size / or_data->scan_size) - 1;
 	if (i < 0) {
-		g_debug ("Not enough data to read (read_size: %d scan_size: %d)", data.read_size, or_data->scan_size);
+		g_debug ("Not enough data to read (read_size: %d scan_size: %d)", (int) data.read_size, or_data->scan_size);
 		return or_data->previous_orientation;
 	}
